migrate to mini-libraries
much easier to understand dependencies between parts of the main library
